services/indexer/requirements.txt
google-genai==1.3.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
